services/admin.py
# -*- coding: utf-8 -*-
from __future__ import absolute_import, unicode_literals

from django import forms
from django.contrib import admin
from modeltranslation.admin import TranslationAdmin

from .models import Service, Category

@admin.register(Service)
class ServiceAdmin(admin.ModelAdmin):
    pass

@admin.register(Category)
class CategoryAdmin(admin.ModelAdmin):
    pass
