requirements/local.txt
# Local development dependencies go here
-r base.txt
django-coverage-plugin==1.5.0
coverage==4.3.4

django-extensions==1.7.8
django-debug-toolbar==1.8
Sphinx==1.6.3
Werkzeug==0.12.2
django-test-plus==1.0.18

# improved REPL
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.8.0
